+++ packages/create-sst/bin/presets/examples/api-sst-auth-facebook/templates/packages/functions/src/auth.ts
@@ -40,9 +40,7 @@ export const handler = AuthHandler({
         );
 
         return Session.parameter({
-          redirect: process.env.IS_LOCAL
-            ? "http://127.0.0.1:5173"
-            : StaticSite.site.url,
+          redirect: StaticSite.site.url || "http://127.0.0.1:5173",
           type: "user",
           properties: {
             userID: claims.sub,