+++ packages/sst/src/constructs/NextjsSite.ts
@@ -13,6 +13,7 @@ import { EdgeFunction } from "./EdgeFunction.js";
 import { SsrSite, SsrSiteProps } from "./SsrSite.js";
 import { Size, toCdkSize } from "./util/size.js";
 import { Duration } from "./util/duration.js";
+import { Policy, PolicyStatement } from "aws-cdk-lib/aws-iam";
 
 const __dirname = url.fileURLToPath(new URL(".", import.meta.url));
 
@@ -104,6 +105,12 @@ export class NextjsSite extends SsrSite {
       environment: {
         BUCKET_NAME: this.cdk.bucket.bucketName,
       },
+      initialPolicy: [
+        new PolicyStatement({
+          actions: ["s3:GetObject"],
+          resources: [this.cdk.bucket.arnForObjects("*")],
+        }),
+      ],
     });
   }
 
@@ -286,7 +293,11 @@ export class NextjsSite extends SsrSite {
         "x-op-middleware-request-headers",
         "x-op-middleware-response-headers",
         "x-nextjs-data",
-        "x-middleware-prefetch"
+        "x-middleware-prefetch",
+        // required by server request (in-place routing)
+        "rsc",
+        "next-router-prefetch",
+        "next-router-state-tree"
       ),
       cookieBehavior: cloudfront.CacheCookieBehavior.all(),
       defaultTtl: CdkDuration.days(0),