+++ packages/sst/test/constructs/nextjs-site/.open-next/assets/_next/static/6R79WHzCDSGceMjDHW-8j/_buildManifest.js
@@ -1,0 +1,1 @@
+self.__BUILD_MANIFEST={__rewrites:{beforeFiles:[],afterFiles:[],fallback:[]},"/":["static/css/1beb14451460885a.css","static/chunks/pages/index-582c16a7cdce70ec.js"],"/_error":["static/chunks/pages/_error-8353112a01355ec2.js"],sortedPages:["/","/_app","/_error"]},self.__BUILD_MANIFEST_CB&&self.__BUILD_MANIFEST_CB();